Agentic-AI/Agent-1/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
Pillow==10.1.0
requests==2.31.0
python-json-logger==2.0.7
gunicorn==21.2.0
google-generativeai==0.3.2
google-cloud-firestore==2.13.1
google-auth==2.23.4
python-dotenv==1.0.0
